Updated import for mesh
git-svn-id: svn+ssh://cxserver.sintef.no/svn/Repository/CustusX3@601 bf23512e-66b2-487e-abc1-aab7af62561f

Former-commit-id: 6ff075c7f79410a57aab8f69a27c8520552e3252

+++ source/3DUSReconstruction/cxImport2DUS.cpp
@@ -90,7 +90,7 @@ void Import2DUS::readInputFile(const char *filename)
 {
   this->mFileName = filename;
   // Remove file ending
-  unsigned int loc = mFileName.rfind(".mha");
+  size_t loc = mFileName.rfind(".mha");
   if ( loc == std::string::npos )
     loc = mFileName.rfind(".mhd");
   if ( loc != std::string::npos )
